modify settings to be seperated
