+++ App.js
@@ -1,34 +1,18 @@
-import React, { createContext, useContext, useState } from "react";
-import { Button, StyleSheet, Text, View } from "react-native";
+import React from "react";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { createStackNavigator } from "@react-navigation/stack";
 import { NavigationContainer } from "@react-navigation/native";
 import { Ionicons } from "@expo/vector-icons";
+import { AppProvider } from "./AppContext";
 import Activities from "./Screens/Activities";
 import Diet from "./Screens/Diet";
-import { AppContext, AppProvider } from "./AppContext";
+import Settings from "./Screens/Settings";
 import AddActivity from "./Screens/AddActivity";
 import AddDietEntry from "./Screens/AddDietEntry";
 
 const Tab = createBottomTabNavigator();
 const Stack = createStackNavigator();
 
-const SettingsScreen = () => {
-  const { toggleTheme, theme } = useContext(AppContext);
-  return (
-    <View
-      style={[
-        styles.container,
-        { backgroundColor: theme === "light" ? "#E6E6FA" : "#333" },
-      ]}
-    >
-      <Text style={{ color: theme === "light" ? "#000" : "#fff" }}>
-        Settings
-      </Text>
-      <Button title="Toggle Theme" onPress={toggleTheme} />
-    </View>
-  );
-};
 export default function App() {
   return (
     <AppProvider>
@@ -84,7 +68,7 @@ export default function App() {
                 />
                 <Tab.Screen
                   name="Settings"
-                  component={SettingsScreen}
+                  component={Settings}
                   options={{
                     title: "Settings",
                     headerStyle: {
@@ -123,12 +107,3 @@ export default function App() {
     </AppProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
